Updated comments
Updated comments

+++ src/recursion/problems/PowerOfTwo.java
@@ -1,12 +1,15 @@
 package src.recursion.problems;
 
+/**
+ * Given a power P, find 2^P using recursion
+ */
 public class PowerOfTwo {
     public static void main(String[] args) {
         int n = 6;
         System.out.println(" 2^" + n + " = " + powerOfTwo(n));
     }
 
-    static int powerOfTwo(long n) {
+    static int powerOfTwo(int n) {
         if (n == 0) {
             return 1;
         } else {
